Abstract out builder from stemcell
