--- lib/builders/version.rb
@@ -1,9 +1,0 @@
-module Bosh
-  module Agent
-    VERSION = '0.7.0'
-    BOSH_PROTOCOL = '1'
-    module StemCell
-      VERSION = '0.0.1'
-    end
-  end
-end
