+++ lib/stemcell/builders/noop.rb
