+++ lib/stemcell/builders/redhat.rb
