+++ lib/stemcell/builders/ubuntu.rb
