+++ stemcell.gemspec
@@ -1,7 +1,7 @@
 # -*- encoding: utf-8 -*-
 $:.unshift(File.join(File.dirname(__FILE__), 'lib'))
 
-require "builders/version"
+require "stemcell/version"
 
 Gem::Specification.new do |gem|
   gem.name                        = "stemcell"
